修改了变量类型

+++ define.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef double Rectangle_a;
+typedef long double Rectangle_a;
